Project/Main/requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
pillow==12.1.0
sqlparse==0.5.5
whitenoise==6.11.0
